chore: move some eth rpc things around again to make room for a new blockchain module
