+++ validation/src/blockchain/eth_rpc/mod.rs
@@ -2,6 +2,4 @@ mod http;
 mod types;
 
 pub use http::HttpEthRpc;
-pub use types::{
-    BlockHash, ChainInclusion, Error, EthProofType, EthTxProofInput, TimeProof, TxHash,
-};
+pub use types::{BlockHash, EthProofType, EthTxProofInput, TxHash};
