+++ validation/src/blockchain/mod.rs
@@ -1,2 +1,6 @@
+mod types;
+
 /// The ethereum RPC provider module
 pub mod eth_rpc;
+
+pub use types::{ChainInclusion, TimeProof, Error};
